% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_liblog}
\alias{get_liblog}
\title{get the table of logged library loading calls}
\usage{
get_liblog(
  duplicates = FALSE,
  default = base,
  base = FALSE,
  recommended = FALSE
)
}
\arguments{
\item{duplicates}{should duplicates be included}

\item{default}{wether to include default packages}

\item{base}{wether to include base packages}

\item{recommended}{wether to include recommended packages}
}
\value{
a data.frame
}
\description{
By default the duplicates, as defined by instances of package namespace loaded
several times at the same timestamp, are removed. Default packages as defined by
\code{getOption("defaultPackages")}, base packages as defined by
\code{rownames(installed.packages(priority="base"))} and recommended packages as
defined by \code{rownames(installed.packages(priority="recommended"))}, are
ignored by default.
}
\examples{
tail(get_liblog())
}
